Refactor App into modules and restore animation/virtual hint guards
  - split app logic into app/ modules (navigation, files, file_panel, playback, etc.)
  - add playback module + helpers and clean imports
  - restore animation-frame override and virtual hint suppression after refactor
